+++ src/Context.php
@@ -197,6 +197,10 @@ class Context
 
     /**
      * Load config
+     *
+     * @template T of Config
+     * @param string|class-string<T> $name
+     * @return ($name is class-string<T> ? T : Config)
      */
     public function load(string $name): Config
     {
@@ -204,8 +208,8 @@ class Context
             return $this->configs[$name];
         }
 
-        $resolveName = explode('#', $name)[0];
-        $configClass = Archetype::resolve(Config::class, $resolveName);
+        $configClass = Archetype::resolve(Config::class, $name);
+        $name = $configClass::getRepositoryName();
 
         $manifest = $this->getFinder()->findConfig($name);
         $loader = $this->getLoaderFor($manifest);